app/pyproject.toml
[project]
name = "household-droid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "llama-index>=0.13.0",
    "llama-index-embeddings-ollama>=0.7.0,!=0.8.0",
    "llama-index-llms-ollama>=0.7.0",
    "llama-index-observability-otel>=0.2.0",
    "llama-index-vector-stores-postgres>=0.6.2",
    "python-jose>=3.5.0",
]

[dependency-groups]
dev = ["httpx>=0.28.1", "pytest>=8.4.1", "ruff>=0.12.9"]

[tool.pytest.ini_options]
pythonpath = "."
